+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 ## 2.4.0 (Unreleased)
 IMPROVEMENTS:
 
-* Change resource handling to use locking mechanism when resource parallel handling is not supported by vCD. [GH-#255] 
+* Change resource handling to use locking mechanism when resource parallel handling is not supported by vCD. [GH-255] 
+
+FEATURES:
+
+* **New Resource:** Load Balancer Service Monitor `vcd_lb_service_monitor` - [GH-256]
+* **New Data Source:** Load Balancer Service Monitor `vcd_lb_service_monitor` - [GH-256]
 
 ## 2.3.0 (May 29, 2019)